Fix one error in the Reistrar state machine (wasn't going to LV state at right time).
Make the stream structures share common members

Make the SRP message matching use the TX stream Id instead of assuming that it
matches the MAC address (which it won't if we are not the original talker but
are still sending an SRP message for that stream) - does this ever happen?

initialise the Attribute structures with UNUSED and a type even for the unmatched
MVRP and MMRP entries - so that they can be correctly checked for a match later.