Fix for bugs 15006 and 15007
The logic for checking if bandwidth is already reserved on a particular
port on Listener join was checking the wrong port and hence never
transmitting on it.

The MAP leave code was incorrectly being executed because it did not
check if this node is the Talker and was decreasing the port bandwidth
with the extra byte added. Hence the negative bandwidth.